zenvoco-backend/requirements.txt
fastapi
uvicorn
motor
pymongo
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
openai
python-dotenv
